Split stack tests from test_valve.py.
