+++ tests/unit/faucet/test_valve_stack.py
@@ -1,0 +1,625 @@
+#!/usr/bin/env python
+
+"""Unit tests run as PYTHONPATH=../../.. python3 ./test_valve_stack.py."""
+
+# Copyright (C) 2015 Research and Innovation Advanced Network New Zealand Ltd.
+# Copyright (C) 2015--2019 The Contributors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from functools import partial
+import time
+import unittest
+
+from ryu.lib import mac
+from ryu.ofproto import ofproto_v1_3 as ofp
+
+from faucet import valves_manager
+from faucet import valve_of
+
+from valve_test_lib import (
+    BASE_DP1_CONFIG, CONFIG, STACK_CONFIG, ValveTestBases)
+
+
+class ValveStackRootExtLoopProtectTestCase(ValveTestBases.ValveTestSmall):
+
+    CONFIG = """
+dps:
+    s1:
+%s
+        stack:
+            priority: 1
+        interfaces:
+            1:
+                description: p1
+                stack:
+                    dp: s2
+                    port: 1
+            2:
+                description: p2
+                native_vlan: 100
+            3:
+                description: p3
+                native_vlan: 100
+                loop_protect_external: True
+            4:
+                description: p4
+                native_vlan: 100
+                loop_protect_external: True
+    s2:
+        hardware: 'GenericTFM'
+        dp_id: 0x2
+        interfaces:
+            1:
+                description: p1
+                stack:
+                    dp: s1
+                    port: 1
+            2:
+                description: p2
+                native_vlan: 100
+            3:
+                description: p3
+                native_vlan: 100
+                loop_protect_external: True
+            4:
+                description: p4
+                native_vlan: 100
+                loop_protect_external: True
+""" % BASE_DP1_CONFIG
+
+    def setUp(self):
+        self.setup_valve(self.CONFIG)
+        self.set_stack_port_up(1)
+
+    def test_loop_protect(self):
+        mcast_match = {
+            'in_port': 2,
+            'eth_dst': mac.BROADCAST_STR,
+            'vlan_vid': 0,
+            'eth_type': 0x800,
+            'ipv4_dst': '224.0.0.5',
+        }
+        self.assertTrue(
+            self.table.is_output(mcast_match, port=1),
+            msg='mcast packet not flooded to non-root stack')
+        self.assertTrue(
+            self.table.is_output(mcast_match, port=3),
+            msg='mcast packet not flooded locally on root')
+        self.assertFalse(
+            self.table.is_output(mcast_match, port=4),
+            msg='mcast packet multiply flooded externally on root')
+
+
+class ValveStackNonRootExtLoopProtectTestCase(ValveTestBases.ValveTestSmall):
+
+    CONFIG = """
+dps:
+    s1:
+%s
+        interfaces:
+            1:
+                description: p1
+                stack:
+                    dp: s2
+                    port: 1
+            2:
+                description: p2
+                native_vlan: 100
+            3:
+                description: p3
+                native_vlan: 100
+                loop_protect_external: True
+            4:
+                description: p4
+                native_vlan: 100
+                loop_protect_external: True
+    s2:
+        hardware: 'GenericTFM'
+        dp_id: 0x2
+        interfaces:
+            1:
+                description: p1
+                stack:
+                    dp: s1
+                    port: 1
+            2:
+                description: p2
+                stack:
+                    dp: s3
+                    port: 1
+            3:
+                description: p2
+                native_vlan: 100
+    s3:
+        hardware: 'GenericTFM'
+        dp_id: 0x3
+        stack:
+            priority: 1
+        interfaces:
+            1:
+                description: p1
+                stack:
+                    dp: s2
+                    port: 2
+            2:
+                description: p2
+                native_vlan: 100
+""" % BASE_DP1_CONFIG
+
+    def setUp(self):
+        self.setup_valve(self.CONFIG)
+        self.set_stack_port_up(1)
+
+    def test_loop_protect(self):
+        mcast_match = {
+            'in_port': 2,
+            'eth_dst': mac.BROADCAST_STR,
+            'vlan_vid': 0,
+            'eth_type': 0x800,
+            'ipv4_dst': '224.0.0.5',
+        }
+        self.assertTrue(
+            self.table.is_output(mcast_match, port=1),
+            msg='mcast packet not flooded to root of stack')
+        self.assertFalse(
+            self.table.is_output(mcast_match, port=3),
+            msg='mcast packet flooded locally on non-root')
+        self.assertFalse(
+            self.table.is_output(mcast_match, port=4),
+            msg='mcast packet flooded locally on non-root')
+
+
+class ValveStackAndNonStackTestCase(ValveTestBases.ValveTestSmall):
+
+    CONFIG = """
+dps:
+    s1:
+%s
+        stack:
+            priority: 1
+        interfaces:
+            1:
+                description: p1
+                stack:
+                    dp: s2
+                    port: 1
+            2:
+                description: p2
+                native_vlan: 0x100
+    s2:
+        hardware: 'GenericTFM'
+        dp_id: 0x2
+        interfaces:
+            1:
+                description: p1
+                stack:
+                    dp: s1
+                    port: 1
+            2:
+                description: p2
+                native_vlan: 0x100
+    s3:
+        hardware: 'GenericTFM'
+        dp_id: 0x3
+        interfaces:
+            1:
+                description: p1
+                native_vlan: 0x100
+            2:
+                description: p2
+                native_vlan: 0x100
+""" % BASE_DP1_CONFIG
+
+    def setUp(self):
+        self.setup_valve(self.CONFIG)
+
+    def test_nonstack_dp_port(self):
+        self.assertEqual(None, self.valves_manager.valves[0x3].dp.shortest_path_port('s1'))
+
+
+class ValveStackRedundancyTestCase(ValveTestBases.ValveTestSmall):
+    """Valve test for updating the stack graph"""
+
+    CONFIG = STACK_CONFIG
+
+    def setUp(self):
+        self.setup_valve(self.CONFIG)
+
+    def test_redundancy(self):
+        now = 1
+        # All switches are down to start with.
+        for dpid in self.valves_manager.valves:
+            self.valves_manager.valves[dpid].dp.dyn_running = False
+        for valve in self.valves_manager.valves.values():
+            self.assertFalse(valve.dp.dyn_running)
+            self.assertEqual('s1', valve.dp.stack_root_name)
+        # From a cold start - we pick the s1 as root.
+        self.assertEqual(None, self.valves_manager.meta_dp_state.stack_root_name)
+        self.assertFalse(self.valves_manager.maintain_stack_root(now))
+        self.assertEqual('s1', self.valves_manager.meta_dp_state.stack_root_name)
+        self.assertEqual(1, self.get_prom('faucet_stack_root_dpid', bare=True))
+        now += (valves_manager.STACK_ROOT_DOWN_TIME * 2)
+        # Time passes, still no change, s1 is still the root.
+        self.assertFalse(self.valves_manager.maintain_stack_root(now))
+        self.assertEqual('s1', self.valves_manager.meta_dp_state.stack_root_name)
+        self.assertEqual(1, self.get_prom('faucet_stack_root_dpid', bare=True))
+        # s2 has come up, but s1 is still down. We expect s2 to be the new root.
+        self.valves_manager.meta_dp_state.dp_last_live_time['s2'] = now
+        now += (valves_manager.STACK_ROOT_STATE_UPDATE_TIME * 2)
+        self.assertTrue(self.valves_manager.maintain_stack_root(now))
+        self.assertEqual('s2', self.valves_manager.meta_dp_state.stack_root_name)
+        self.assertEqual(2, self.get_prom('faucet_stack_root_dpid', bare=True))
+        # More time passes, s1 is still down, s2 is still the root.
+        now += (valves_manager.STACK_ROOT_DOWN_TIME * 2)
+        # s2 recently said something, s2 still the root.
+        self.valves_manager.meta_dp_state.dp_last_live_time['s2'] = now - 1
+        self.assertFalse(self.valves_manager.maintain_stack_root(now))
+        self.assertEqual('s2', self.valves_manager.meta_dp_state.stack_root_name)
+        self.assertEqual(2, self.get_prom('faucet_stack_root_dpid', bare=True))
+        # now s1 came up too, so we change to s1 because we prefer it.
+        self.valves_manager.meta_dp_state.dp_last_live_time['s1'] = now + 1
+        now += valves_manager.STACK_ROOT_STATE_UPDATE_TIME
+        self.assertTrue(self.valves_manager.maintain_stack_root(now))
+        self.assertEqual('s1', self.valves_manager.meta_dp_state.stack_root_name)
+        self.assertEqual(1, self.get_prom('faucet_stack_root_dpid', bare=True))
+
+
+class ValveRootStackTestCase(ValveTestBases.ValveTestSmall):
+    """Test stacking/forwarding."""
+
+    DP = 's3'
+    DP_ID = 0x3
+
+    def setUp(self):
+        self.setup_valve(CONFIG)
+        self.set_stack_port_up(5)
+
+    def test_stack_learn(self):
+        """Test host learning on stack root."""
+        self.prom_inc(
+            partial(self.rcv_packet, 1, 0x300, {
+                'eth_src': self.P1_V300_MAC,
+                'eth_dst': self.UNKNOWN_MAC,
+                'ipv4_src': '10.0.0.1',
+                'ipv4_dst': '10.0.0.2'}),
+            'vlan_hosts_learned',
+            labels={'vlan': str(int(0x300))})
+
+    def test_stack_flood(self):
+        """Test packet flooding when stacking."""
+        matches = [
+            {
+                'in_port': 1,
+                'vlan_vid': 0,
+                'eth_src': self.P1_V300_MAC
+            }]
+        self.verify_flooding(matches)
+
+
+class ValveEdgeStackTestCase(ValveTestBases.ValveTestSmall):
+    """Test stacking/forwarding."""
+
+    DP = 's4'
+    DP_ID = 0x4
+
+    def setUp(self):
+        self.setup_valve(CONFIG)
+        self.set_stack_port_up(5)
+
+    def test_stack_learn(self):
+        """Test host learning on non-root switch."""
+        self.rcv_packet(1, 0x300, {
+            'eth_src': self.P1_V300_MAC,
+            'eth_dst': self.UNKNOWN_MAC,
+            'ipv4_src': '10.0.0.1',
+            'ipv4_dst': '10.0.0.2'})
+        self.rcv_packet(5, 0x300, {
+            'eth_src': self.P1_V300_MAC,
+            'eth_dst': self.UNKNOWN_MAC,
+            'vid': 0x300,
+            'ipv4_src': '10.0.0.1',
+            'ipv4_dst': '10.0.0.2'})
+
+    def test_stack_flood(self):
+        """Test packet flooding when stacking."""
+        matches = [
+            {
+                'in_port': 1,
+                'vlan_vid': 0,
+                'eth_src': self.P1_V300_MAC
+            }]
+        self.verify_flooding(matches)
+
+    def test_no_unexpressed_packetin(self):
+        """Test host learning on stack root."""
+        unexpressed_vid = 0x666 | ofp.OFPVID_PRESENT
+        match = {
+            'vlan_vid': unexpressed_vid,
+            'eth_dst': self.UNKNOWN_MAC}
+        self.assertFalse(
+            self.table.is_output(match, port=ofp.OFPP_CONTROLLER, vid=unexpressed_vid))
+
+
+class ValveStackProbeTestCase(ValveTestBases.ValveTestSmall):
+    """Test stack link probing."""
+
+    CONFIG = STACK_CONFIG
+
+    def setUp(self):
+        self.setup_valve(self.CONFIG)
+
+    def test_stack_probe(self):
+        """Test probing works correctly."""
+        stack_port = self.valve.dp.ports[1]
+        other_dp = self.valves_manager.valves[2].dp
+        other_port = other_dp.ports[1]
+        other_valves = self.valves_manager._other_running_valves(self.valve)
+        self.valve.fast_state_expire(time.time(), other_valves)
+        self.assertTrue(stack_port.is_stack_init())
+        for change_func, check_func in [
+                ('stack_up', 'is_stack_up'),
+                ('stack_down', 'is_stack_down')]:
+            getattr(other_port, change_func)()
+            self.rcv_lldp(stack_port, other_dp, other_port)
+            self.assertTrue(getattr(stack_port, check_func)())
+
+    def test_stack_miscabling(self):
+        """Test probing stack with miscabling."""
+        stack_port = self.valve.dp.ports[1]
+        other_dp = self.valves_manager.valves[2].dp
+        other_port = other_dp.ports[1]
+        wrong_port = other_dp.ports[2]
+        wrong_dp = self.valves_manager.valves[3].dp
+        other_valves = self.valves_manager._other_running_valves(self.valve)
+        self.valve.fast_state_expire(time.time(), other_valves)
+        for remote_dp, remote_port in [
+                (wrong_dp, other_port),
+                (other_dp, wrong_port)]:
+            self.rcv_lldp(stack_port, other_dp, other_port)
+            self.assertTrue(stack_port.is_stack_down() or stack_port.is_stack_init())
+            self.rcv_lldp(stack_port, remote_dp, remote_port)
+            self.assertTrue(stack_port.is_stack_down())
+
+    def test_stack_lost_lldp(self):
+        """Test stacking when LLDP packets get dropped"""
+        stack_port = self.valve.dp.ports[1]
+        other_dp = self.valves_manager.valves[2].dp
+        other_port = other_dp.ports[1]
+        other_valves = self.valves_manager._other_running_valves(self.valve)
+        self.valve.fast_state_expire(time.time(), other_valves)
+        self.rcv_lldp(stack_port, other_dp, other_port)
+        self.assertTrue(stack_port.is_stack_init())
+        # simulate packet loss
+        self.valve.fast_state_expire(time.time() + 300, other_valves)
+        self.assertTrue(stack_port.is_stack_down())
+
+
+class ValveStackGraphUpdateTestCase(ValveTestBases.ValveTestSmall):
+    """Valve test for updating the stack graph."""
+
+    CONFIG = STACK_CONFIG
+
+    def setUp(self):
+        self.setup_valve(self.CONFIG)
+
+    def test_update_stack_graph(self):
+        """Test stack graph port UP and DOWN updates"""
+
+        def all_stack_up():
+            for valve in self.valves_manager.valves.values():
+                valve.dp.dyn_running = True
+                for port in valve.dp.stack_ports:
+                    port.stack_up()
+
+        def up_stack_port(port):
+            peer_dp = port.stack['dp']
+            peer_port = port.stack['port']
+            for state_func in [peer_port.stack_init, peer_port.stack_up]:
+                state_func()
+                self.rcv_lldp(port, peer_dp, peer_port)
+            self.assertTrue(port.is_stack_up())
+
+        def down_stack_port(port):
+            up_stack_port(port)
+            peer_port = port.stack['port']
+            peer_port.stack_down()
+            self.valves_manager.valve_flow_services(
+                time.time() + 600,
+                'fast_state_expire')
+            self.assertTrue(port.is_stack_down())
+
+        def verify_stack_learn_edges(num_edges, edge=None, test_func=None):
+            for dpid in (1, 2, 3):
+                valve = self.valves_manager.valves[dpid]
+                if not valve.dp.stack:
+                    continue
+                graph = valve.dp.stack['graph']
+                self.assertEqual(num_edges, len(graph.edges()))
+                if test_func and edge:
+                    test_func(edge in graph.edges(keys=True))
+
+        num_edges = 3
+        all_stack_up()
+        verify_stack_learn_edges(num_edges)
+        ports = [self.valve.dp.ports[1], self.valve.dp.ports[2]]
+        edges = [('s1', 's2', 's1:1-s2:1'), ('s1', 's2', 's1:2-s2:2')]
+        for port, edge in zip(ports, edges):
+            num_edges -= 1
+            down_stack_port(port)
+            verify_stack_learn_edges(num_edges, edge, self.assertFalse)
+        up_stack_port(ports[0])
+        verify_stack_learn_edges(2, edges[0], self.assertTrue)
+
+
+class ValveTestIPV4StackedRouting(ValveTestBases.ValveTestStackedRouting):
+    """Test inter-vlan routing with stacking capabilities in an IPV4 network"""
+
+    VLAN100_FAUCET_VIPS = '10.0.1.254'
+    VLAN100_FAUCET_VIP_SPACE = '10.0.1.254/24'
+    VLAN200_FAUCET_VIPS = '10.0.2.254'
+    VLAN200_FAUCET_VIP_SPACE = '10.0.2.254/24'
+
+    def setUp(self):
+        self.setup_stack_routing()
+
+
+class ValveTestIPV4StackedRoutingDPOneVLAN(ValveTestBases.ValveTestStackedRouting):
+    """Test stacked intervlan routing when each DP has only one of the routed VLANs"""
+
+    VLAN100_FAUCET_VIPS = '10.0.1.254'
+    VLAN100_FAUCET_VIP_SPACE = '10.0.1.254/24'
+    VLAN200_FAUCET_VIPS = '10.0.2.254'
+    VLAN200_FAUCET_VIP_SPACE = '10.0.2.254/24'
+
+    def base_config(self):
+        """Create the base config"""
+        self.V100_HOSTS = [1]
+        self.V200_HOSTS = [2]
+        return """
+    routers:
+        router1:
+            vlans: [vlan100, vlan200]
+    dps:
+        s1:
+            hardware: 'GenericTFM'
+            dp_id: 1
+            stack: {priority: 1}
+            interfaces:
+                1:
+                    native_vlan: vlan100
+                3:
+                    stack: {dp: s2, port: 3}
+        s2:
+            dp_id: 2
+            interfaces:
+                2:
+                    native_vlan: vlan200
+                3:
+                    stack: {dp: s1, port: 3}
+    """
+
+    def setUp(self):
+        self.setup_stack_routing()
+
+
+class ValveTestIPV4StackedRoutingPathNoVLANS(ValveTestBases.ValveTestStackedRouting):
+    """Test stacked intervlan routing when DP in path contains no routed VLANs"""
+
+    VLAN100_FAUCET_VIPS = '10.0.1.254'
+    VLAN100_FAUCET_VIP_SPACE = '10.0.1.254/24'
+    VLAN200_FAUCET_VIPS = '10.0.2.254'
+    VLAN200_FAUCET_VIP_SPACE = '10.0.2.254/24'
+
+    def create_config(self):
+        """Create the config file"""
+        self.CONFIG = """
+    vlans:
+        vlan100:
+            vid: 0x100
+            faucet_mac: '%s'
+            faucet_vips: ['%s']
+        vlan200:
+            vid: 0x200
+            faucet_mac: '%s'
+            faucet_vips: ['%s']
+        vlan300:
+            vid: 0x300
+    %s
+           """ % (self.VLAN100_FAUCET_MAC, self.VLAN100_FAUCET_VIP_SPACE,
+                  self.VLAN200_FAUCET_MAC, self.VLAN200_FAUCET_VIP_SPACE,
+                  self.base_config())
+
+    def base_config(self):
+        """Create the base config"""
+        self.V100_HOSTS = [1]
+        self.V200_HOSTS = [3]
+        return """
+    routers:
+        router1:
+            vlans: [vlan100, vlan200]
+    dps:
+        s1:
+            hardware: 'GenericTFM'
+            dp_id: 1
+            stack: {priority: 1}
+            interfaces:
+                1:
+                    native_vlan: vlan100
+                3:
+                    stack: {dp: s2, port: 3}
+        s2:
+            dp_id: 2
+            interfaces:
+                2:
+                    native_vlan: vlan300
+                3:
+                    stack: {dp: s1, port: 3}
+                4:
+                    stack: {dp: s3, port: 3}
+        s3:
+            dp_id: 3
+            interfaces:
+                2:
+                    native_vlan: vlan200
+                3:
+                    stack: {dp: s2, port: 4}
+                4:
+                    stack: {dp: s4, port: 3}
+        s4:
+            dp_id: 4
+            interfaces:
+                2:
+                    native_vlan: vlan300
+                3:
+                    stack: {dp: s3, port: 4}
+    """
+
+    def setUp(self):
+        self.setup_stack_routing()
+
+
+class ValveTestIPV6StackedRouting(ValveTestBases.ValveTestStackedRouting):
+    """Test inter-vlan routing with stacking capabilities in an IPV6 network"""
+
+    VLAN100_FAUCET_VIPS = 'fc80::1:254'
+    VLAN200_FAUCET_VIPS = 'fc80::2:254'
+    VLAN100_FAUCET_VIP_SPACE = 'fc80::1:254/64'
+    VLAN200_FAUCET_VIP_SPACE = 'fc80::1:254/64'
+
+    def setUp(self):
+        self.setup_stack_routing()
+
+    @staticmethod
+    def create_ip(vindex, host):
+        """Create a IP address string"""
+        return 'fc80::%u:%u' % (vindex, host)
+
+    @staticmethod
+    def get_eth_type():
+        """Returns IPV6 ether type"""
+        return valve_of.ether.ETH_TYPE_IPV6
+
+    def create_match(self, vindex, host, faucet_mac, faucet_vip, code):
+        """Create an NA message"""
+        return {
+            'eth_src': self.create_mac(vindex, host),
+            'eth_dst': faucet_mac,
+            'ipv6_src': self.create_ip(vindex, host),
+            'ipv6_dst': faucet_vip,
+            'neighbor_advert_ip': self.create_ip(vindex, host)
+        }
+
+
+if __name__ == "__main__":
+    unittest.main()  # pytype: disable=module-attr
